FIX(perf): Use `new_unchecked()` when appropriate.
Objects can be constructed by the nom parser or through the API. Both
interfaces need to check byte sequences (or strings) for conformance.
However, when nom already ensures that a byte sequence conforms to IMAP,
a second check is wasteful. For now, and if there is not another clever
method, we introduced `new_unchecked()` to just bypass this redundant check.