feat: Add Python bindings for decoders
Wrap codecs from `imap-codec` into `pyobject` types to be used from
Python.
For each of the codecs, the `decode` method is implemented and
corresponding Python tests are added to prevent regressions on the
bindings. The encoding implementation will be added in a subsequent
commit.
Currently, the resulting `dict`/`tuple` types are still very rough which
should be changed further down the road.