build: Add CI workflow for publishing Rust crates
This adds a basic CI workflow for automatically publishing the Rust
crates (currently `imap-codec` and `imap-types`) to crates.io when a
corresponding Git tag is pushed.
For now, only the tag's version number is checked against the version
number specified in the crates `Cargo.toml` file before publishing the
crate is performed.